Update README.md
Welcome to the repository for "My First JavaScript Portfolio Website Build (2019)". This project marks a significant milestone in my journey as a web developer, representing my initial foray into creating a professional portfolio website. Built in 2019, this site was designed to highlight my skills and projects using a modern tech stack, including JavaScript, Tailwind CSS, Next.js, and clsx.

The primary goal of this portfolio website is to provide a platform where I can showcase my work, demonstrate my capabilities in web development, and illustrate my growth in the field. Through this project, I aimed to create a visually appealing and highly functional website that reflects contemporary design principles and best practices in web development.

The site features a responsive design, ensuring optimal viewing experiences across various devices and screen sizes. By leveraging the power of Tailwind CSS, I was able to implement a utility-first approach to styling, resulting in a clean and consistent user interface. The use of Next.js for server-side rendering and static site generation enhances the performance and scalability of the website, making it a robust solution for presenting my work.

Overall, this project not only showcases my technical skills and knowledge but also my dedication to continuous learning and improvement in the ever-evolving landscape of web development. I hope this portfolio serves as a testament to my capabilities and a stepping stone for future projects and opportunities